AutoresMejorado/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
flask-bcrypt==1.0.1
PyJWT==2.8.0